KOGTIO-5804: prefix package as org.kie.kogito - dmn-listener[quarkus|springboot]
