+++ dmn-listener-quarkus/src/test/java/org/kie/kogito/dmn/quarkus/example/mock/MockDMNRuntimeEventListener.java
